refactor(libs): move error libs to the same file
