+++ service/src/app/controllers/movie.js
@@ -1,5 +1,5 @@
 import movieModel from '../models/movie'
-import catchResponse from '../libs/catchResponse'
+import { catchResponse } from '../libs/resErrorHandling'
 import ERROR_RESPONSE from '../constants/errorResponse'
 
 const getAll = async (req, res) => {
