--- service/src/app/libs/errorMiddleware.js
@@ -1,1 +1,0 @@
-export default (err, req, res, _next) => res.status(err.httpStatus).json(err)
