+++ service/src/server.js
@@ -4,7 +4,7 @@ import bodyParser from 'body-parser'
 
 import database from './app/connections/database'
 import router from './app/routes'
-import errorResponse from './app/libs/errorMiddleware'
+import { errorMiddleware } from './app/libs/resErrorHandling'
 
 const app = express()
 const port = process.env.SERVER_PORT || 3002
@@ -16,7 +16,7 @@ app.use(bodyParser.json({ limit: '20mb' }))
 app.use(bodyParser.json({ type: 'application/vnd.api+json' }))
 app.use(bodyParser.urlencoded({ limit: '20mb', extended: true }))
 app.use(process.env.URL_PREFIX, router)
-app.use(errorResponse)
+app.use(errorMiddleware)
 
 app.get('/healthz', (req, res) => res.status(200).json({ status: 'ok' }))
